Clean generic impls code
